Added support for static input-numbers and expressions
Summary:
Mobile can now trigger a `staticRender` option in the Perseus API to have input-number
and expression widgets render as pretty latex (which will be set through the Perseus API). The
latex is rendered with the <Tex> component, which is then combined with some styling and wrapped
into a <MathOutput> component.

For some reason, this doesn't work when answer-area != "multiple", but I saw some other issues
related to that on Trello, so maybe it's not my fault?

Test Plan:
- Run the mobile app.
- Try "Groups of Tens" to test out input-number.
- To verify that input-number is grading mixed numbers correctly (but displays them as Tex), try out: "Adding and Subtracting Fractions with Like Denominators Word Problems" (some of the items have answer-area != "multiple", so those won't work).
- Try "Writing Expressions with Exponents" to test out expression.

Reviewers: jack

Differential Revision: http://phabricator.khanacademy.org/D11073