Break out of exponent on *, /, and - in the expression editor
Summary: We want to break out the exponent in the expression editor on  `*`, `/`, and `-` except when the `-` is being used for a negative exponent. Fixes T2730 and merges in changes from D11366 changes in MathQuill.

Test Plan:
Add an expression editor widget
Verify that when you type `2^3*4` it becomes `(2^3)*4`
Verify that when you type `2^3/4` it becomes `(2^3)/4`
Verify that when you type `2^3-4` it becomes `(2^3)-4`
Verify that when you type `2^-3` it does not break out of the exponent on `-`.

Reviewers: jack

Reviewed By: jack

Subscribers: jack, joel, alex, alpert, eater

Maniphest Tasks: T2730

Differential Revision: http://phabricator.khanacademy.org/D11585