Add beginnings of a parser.
Summary:
I'm using jison somewhat arbitrarily but it seems to have more uptake
than peg.js.

It's worth noting that I wrote
[jison-loader](https://github.com/joelburget/jison-loader), but it's
trivial.

At the moment this thing can successfully parse "5 m^2" but not "5.5 m"
or "5 m/s".

One more note - the "browser: { fs: false }" thing is because the
jison-generated code includes a require of fs, which blows up unless we
tell webpack to ignore it.

Test Plan:
Successfully load the perseus test page and attempt to parse the examples above
with `p`.

Reviewers: alex

Subscribers: aria

Differential Revision: http://phabricator.khanacademy.org/D14288