Util: Add a function to relocate an image to our servers
Summary:
In addition to uploading data from one's computer, one might
want to relocate an image that is already online. This is
one of the actions available in the ImageUploadDialog
perseus one uses, and which the image widget will soon
use.

In perseus one, we do this with a private API endpoint,
but since I don't want to either add private api endpoint
code into perseus or pass down a second relocateImage
API option, I've instead written a js wrapper that can
relocate an image by downloading it and then uploading
it with `apiOptions.uploadImage`.

Disclaimer: Please only do this in the case of public domain
or CC images that are legal to use!

Test Plan:
Tested in the image uploading final commit, where I:
 * transfered an svg image
 * transfered a jpeg image
 * transfered a png image

Reviewers: alex