Image upload: Copy-pasta the image upload dialog from perseus one
Summary:
I want to add image uploading to the image widget. This isn't the
right way to solve this long-term, but having this dialog as
a part of perseus proper makes some sense because this isn't
tied to the server and is something that makes sense to have
in more places in perseus with the refactoring of
`apiOptions.uploadImage` being available in more places.

There is actually a lot of duplication in the logic around
uploading images already, mostly around content parsing
and what to do with replacing image placeholders
(see the editor and perseus one item-area). We should
spend some time thinking more deeply about how to
properly unify this, as well as making this dialog
nicer, in the future.

For now, since we don't have a good way of sharing
files between perseus-one/webapp and perseus,
and it would not make sense to pass this component
down to perseus from webapp (imho), I think this
is not great but not horrible, hopefully.

NOTE: This diff *just* literally copies the files from
webapp. They don't quite work in perseus yet, but I
will fix that in the next commit, so that it is easy
to see the changes that I made.

Test Plan:
Tested as a part of the image widget's image upload
changes.

Reviewers: alex