Added double columns and inter widget communication
Unfortunately to properly get widgets to communicate
through the current implementation of interWidgets we had
to implement a two pass render. It would be nice to
reimplement interWidgets communication so there is a cleaner
way to do this, but we are not going to do that now.

Test Plan:
- open /multirenderer.html
- Go to the bottom and see that the radio buttons reference the passage

Auditors: johnsullivan