Handle changing item data in the MultiRenderer.
Summary:
This change addresses two Trello cards:

* https://trello.com/c/Cvhniyc5
* https://trello.com/c/cSYfIW0c

First some background: the MultiRenderer requires two passes to render
properly. First it renders the context (left column), then it renders
the questions (right column). This is due to specifics of inter widget
communication.

Previously we weren't properly dealing with the case of our questions'
or context's item data changing after the MultiRenderer was mounted. We
should re-render the context, and then render the questions. Instead we
were just rendering everything all at once.

Additionally, we were reusing the Renderer components where possible
(because React tries to do that when it can). There have been issues in
the past around the Perseus Renderer not wiping all of the state it
should and things getting a little weird.

I'm not sure whether that's still a problem (maybe @aria knows), but we
can handle it fairly cleanly here (and in such a way that it's easy to
stop doing it in the future if we're sure that it's safe to reuse
Renderers), so here we are.

Test Plan:
A.1) Go to http://localhost:9000/multirenderer.html
A.2) Interact with one of the widgets such that its state changes (fill
     in a text box, click a radio button, w/e).
A.3) Run `switchItem(true)` in your console. See the question numbers
     change, but the change you made to the widget's state should still
     be visible.

B.1) Go to http://localhost:9000/multirenderer.html
B.2) Interact with one of the widgets such that its state changes (fill
     in a text box, click a radio button, w/e).
B.3) Run `switchItem(false)` in your console. See the question numbers
     change and the order of two questions change. The change you made
     to the widget's state should be gone and the widget should be
     reset to its original state.

Reviewers: phillip

Reviewed By: phillip

Subscribers: kevindangoor, aria

Differential Revision: https://phabricator.khanacademy.org/D19795