Created a multirenderer editor and added it to multirenderer.html
Summary:
This is the beginning of work being done for this trello card https://trello.com/c/Ztbsdrod/25-start-smashing-on-the-editor-ui

In this diff I have added a third column to multirenderer.html that contains editors for the questions passed into the Multirenderer. The passage editor has not yet been added. Below is a screenshot of what multirenderer.html looks like after this diff.

{F162848}

Test Plan:
- Start perseus server
- Go to localhost:9000/multirenderer.html
- See three columns with editors, a passage, and questions

Reviewers: johnsullivan

Reviewed By: johnsullivan

Subscribers: aria, kevindangoor

Differential Revision: https://phabricator.khanacademy.org/D19937