Change item data. json->questions.
Summary:
There are a few misc things that were done in this change. There is no
corresponding Trello card :(.

Multi item data now has this shape:

    var multiItemDataShape = PT.shape({
        context: PT.shape({
            content: PT.string,
            widgets: PT.object,
            images: PT.object,
        }),

        questions: PT.arrayOf(PT.shape({
            content: PT.string,
            widgets: PT.object,
            images: PT.object,
            hints: PT.array.isRequired,
        })).isRequired,
    });

I've added an example item to multirenderer.html that uses this shape.
This is what we'll store in the datastore and will hopefully serve us
well.

I also cleaned up the editor props a bit to make things clearer. The
editor now accepts a questions prop that behaves identically to the
renderer's questions prop. This was basically a find and replace of
this.props.json to this.props.questions.

Test Plan:
1) Go to /multirenderer.html
2) See an item appear (a cool passage and two questions)
3) Modify one of the questions
4) See the preview update itself

Reviewers: phillip

Reviewed By: phillip

Subscribers: aria, kevindangoor

Differential Revision: https://phabricator.khanacademy.org/D20310