Add serialized state support to multirenderer.
Summary:
This is part of a larger effort to add support for all of the features
the SAT mission requires into the MultiRenderer. Trello card:
https://trello.com/c/9LGQe0Ku

Here we add support for serialized state. I've created an interface
that is different than the interface that the Renderer currently
exposes. Namely, when getting and restoring serialized state for the
Renderer, you need to do something like:

    var state = renderer.getSerializedState();
    renderer.restoreSerializedState(state);

Because the Renderer is not always ready to recieve serialized state
after a single render (I think because when group widgets are around,
we need multiple renders), we had to do creative things to send down
the serialized state only after some amount of renders. This is error
prone and caused a lot of pain during the SAT mission's development.

The interface I've come up with for the MultiRenderer involves
passing the serialized state to the MultiRenderer in a prop, as well as
a "problemId". When the MultiRenderer first finishes mounting (which
actually requires two renders), it will restore that state, as well as
whenever the problemId changes.

I plan to change the interface of the Renderer to match the
MultiRenderer's interface in the future (I will write an internal
project doc or support ticket for this, but have not yet).

{F167660}

Test Plan:
1) Go to /multirenderer.html
2) Select the second radio option
3) Click "save serialized state"
4) Select the first radio option
5) Click "restore serialized state"
6) See that the second radio option is again selected

Reviewers: phillip

Reviewed By: phillip

Subscribers: aria

Differential Revision: https://phabricator.khanacademy.org/D20352