Flatten columns in Multi-Renderer editor.
Summary:
Trello card: https://trello.com/c/nLqhqu0M

Currently, the Multi-Renderer editor is very wide because there is a
column column for the individual editors, a column for the context, and
a column for the questions.

To allow content creators with normal sized screens to use the editor,
we now flatten the columns and display the questions below the context.

{F170245}

Test Plan:
1) GO to /multirenderer.html
2) See that the questions are rendered beneath the context

Reviewers: phillip

Reviewed By: phillip

Subscribers: aria

Differential Revision: https://phabricator.khanacademy.org/D20650