Have the MultiRenderer take in item data blob.
Summary:
Trello card: https://trello.com/c/iR7dePu6

Prior to this card, the MultiRenderer would take in the questions and
context separately. This forces the client to actually pull those
things out of the item data which in turn forces us to guarentee that
(1) both those attributes always exist within a major version and (2)
only those attributes are interesting to us within a major version.

We don't have any reason to place those restrictions on ourselves, so
this change gets rid of them in favor of taking the whole item data
blob.

This was prompted by https://trello.com/c/DbmjUDLC, which involves
adding an itemRendererType to the item data. If we had already shipped
and wanted to do this, it would require a change all client code, which
would be unfortunate.

Test Plan:
Confirm that the following behaviors work as expected in the editor at
/multirenderer.html:

* Modifying the contents of the context
* Modifying the contents of a question
* Deleting a question
* Moving a question up
* Moving a question down
* Creating a new question
* Modifying the JSON in Developer JSON Mode
* Permalinking

Reviewers: phillip

Reviewed By: phillip

Subscribers: aria

Differential Revision: https://phabricator.khanacademy.org/D20700