Implement preview mode in MultiRenderer editor.
Summary:
Trello card: https://trello.com/c/DXbeEIcp

This allows users to see a preview of what the item will look like when
rendered in a real two-column multi-renderer. This is especially
important now that D20650 and we don't display two columns in the
live-updating preview available in edit mode.

{F170325}

Test Plan:
1) Go to /multirenderer.html
2) Switch into preview mode
3) Verify things look alright (see screenshot above)
4) Switch into JSON mode
5) Verify things look alright (you should see a large textbox with JSON
   in it).

Reviewers: phillip

Reviewed By: phillip

Subscribers: aria

Differential Revision: https://phabricator.khanacademy.org/D20669