Style Multi-Renderer in SAT product.
Summary:
Trello card: https://trello.com/c/QngP9bxI

This applies the SAT specific styling requirements to the
Multi-Renderer when rendered in the SAT product, without breaking the
styles for the normal Renderer & Group Widgets combo.

Two-Column Multi-Renderer in SAT:

{F172395}

{F172396}

Single-Column Multi-Renderer in SAT:

{F172398}

{F172399}

Old-School Renderer in SAT:

{F172400}

{F172401}

Test Plan:
To test this patch, you'll need to first bring in D20753 to your copy
of webapp, and update your Perseus version on webapp to be up to date
with this patch.

1) Go to /mission/sat in webapp
2) Load up a diagnostic
3) Go through the questions and confirm they look like the screenshots
   above.

Reviewers: kitt

Reviewed By: kitt

Subscribers: aria

Differential Revision: https://phabricator.khanacademy.org/D20849