Update travis config
Summary:
Travis was trying to run this on an older version of node, getting the
error message of:

```
/home/travis/build/Khan/perseus/node_modules/jsdom/lib/jsdom.js:3
() => `jsdom 7.x onward only works on Node.js 4 or newer:
https://github.com/t
 ^
```

This commit tells travis to use node 4.1 to run our tests.

Test plan:
Push this to a remote branch and see if the build passes on travis.
If it does, merge it.

Auditors: michelle