Bundle `math-input` stylesheets with Perseus
Summary:
`math-input` has a couple of stylesheets for animations and overriding some of the MathQuill
defaults. This diff bundles those stylesheets with Perseus so that we won't have to pull them
in separately in webapp.

The variables (which are rendered as text and not SVG) aren't quite right because we don't have
Proxima on that page, but I believe we'll have Proxima just fine in webapp with no additional work.

On Asana @ https://app.asana.com/0/114493820703878/133129670297602.

Test Plan:
- Open up an Expression widget with `customKeypad: true`.
    - Verify that the cursor is green.
    - Verify that the selection rect is green.
- Disable `customKeypad`.
    - Verify that the cursor is back to black.
- Run `make build/perseus.css`.
- Open up `perseus.less`.
    - Verify that it contains styles for the `dismissable` and `echo` classes.

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27325